Add BlockReader
Adds BlockReader, which allows you to read blocks of characters delimited
by a given sequence from an arbitrary input source.

It provides line numbering relative to that source, as well as block
numbering.

The main caveat is that to prevent spurious NoSuchElementExceptions, EOF
is always treated as a valid delimiter, so blocks may be shorter than you
expect them to be.

This is only meant as a data input utility, not a data validation utility.