Rename exceptions to not use Exception
This renames several of the stock exceptions to remove 'Exception' from
their name. This is something that I don't recall where I picked it up
from, but the idea is that the fact that it is an exception is generally
syntactically obvious. Therefore, you can instead give your exceptions
names which directly state what it wrong, as well as strongly implying
what you can do to fix it.

For instance, if you get a DirectionInvalid, it is apparent what the
issue is (it is less apparent why it is invalid, but that is another
issue, and a discussion for another time)