Add unit tests for defines
Adds unit tests for SimpleDefine and IteratedDefine.

This also fixes an issue with IteratedDefine, where once you had
consumed a replacer, it was consumed for good; you couldn't use it in
the future, even in a different call to apply().

This was fixed through the introduction of a new iterator type from
esodata - ResettableIterator. See that project/type for more details on
what exactly this does; but suffice to say, it allows to restore our
iterator and re-iterate over the same elements on every call to apply.